refactor: move dialogporten graphql endpoint into `./src/graphql`
