+++ packages/bff/src/auth/userApi.ts
@@ -21,30 +21,6 @@ interface RequestBody {
 }
 
 const plugin: FastifyPluginAsync = async (fastify, options) => {
-  fastify.post(
-    '/api/test',
-    { preValidation: fastify.verifyToken },
-    async (request: FastifyRequest, reply: FastifyReply) => {
-      try {
-        const token = request.session.get('token');
-        const headers = {
-          'content-type': 'application/json',
-          Authorization: `Bearer ${token!.access_token}`,
-        };
-
-        const response = await axios({
-          method: 'POST',
-          url: 'https://dp-be-test-graphql-ca.lemonisland-e7641482.norwayeast.azurecontainerapps.io/graphql',
-          data: request.body,
-          headers,
-        });
-        reply.send(response.data);
-      } catch (e) {
-        reply.status(500).send({ error: 'Internal Server Error', message: (e as unknown as Error).message });
-      }
-    },
-  );
-
   fastify.get(
     '/api/isAuthenticated',
     { preValidation: fastify.verifyToken },
