+++ packages/bff/src/graphql/api.ts
@@ -1,7 +1,8 @@
 import { FastifyPluginAsync } from 'fastify';
 import fp from 'fastify-plugin';
 import { createHandler } from 'graphql-http/lib/use/fastify';
-import { schema } from './schema.js';
+import config from '../config.ts';
+import { schema } from './schema.ts';
 
 const plugin: FastifyPluginAsync = async (fastify, options) => {
   fastify.post(
@@ -15,6 +16,30 @@ const plugin: FastifyPluginAsync = async (fastify, options) => {
         };
       },
     }),
+  );
+
+  fastify.post(
+    '/api/graphql2',
+    { preValidation: fastify.verifyToken },
+    async (request: FastifyRequest, reply: FastifyReply) => {
+      try {
+        const token = request.session.get('token');
+        const headers = {
+          'content-type': 'application/json',
+          Authorization: `Bearer ${token!.access_token}`,
+        };
+
+        const response = await axios({
+          method: 'POST',
+          url: config.dialogportenURL + '/graphql',
+          data: request.body,
+          headers,
+        });
+        reply.send(response.data);
+      } catch (e) {
+        reply.status(500).send({ error: 'Internal Server Error', message: (e as unknown as Error).message });
+      }
+    },
   );
 };
